pyproject.toml
[project]
name = "fracfocus-data-api"
version = "1.0.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.30.6",
    "sqlalchemy==2.0.36",
"requests==2.32.3",
    "pydantic-settings==2.5.2",
    "apscheduler==3.10.4",
    "python-json-logger==2.0.7",
]

[project.optional-dependencies]
dev = [
    "httpx==0.27.2",
    "pytest==8.3.3",
    "pytest-mock==3.14.0",
]

[dependency-groups]
dev = [
    "httpx==0.27.2",
    "pytest==8.3.3",
    "pytest-mock==3.14.0",
]
